:building_construction: Retrieveng home has been changed from home page to app component
